*refactoring of ConvertUnsqueezeNode.
